+++ packages/components/src/components/badge/badge.figma.ts
@@ -1,0 +1,17 @@
+import figma, { html } from '@figma/code-connect';
+
+figma.connect(
+  'https://www.figma.com/design/mfiAVMWkxiBRGnegjqLMNW/MiniS-DS?node-id=4605-441',
+  {
+    props: {
+      color: figma.enum('Color', {
+        Pink: 'pink',
+        Yellow: 'yellow',
+        Blue: 'blue',
+      }),
+    },
+    example: ({ color }) => html`
+      <minis-badge color="${color}"></minis-badge>
+    `,
+  },
+);